requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.0.1
django-cors-headers==4.3.1
python-decouple==3.8
reportlab==4.0.7